requirements.txt
asgiref==3.7.2
attrs==23.1.0
base58==2.1.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorharmonies==1.0.5
colour==0.1.5
contourpy==1.0.5
CTkMessagebox==2.1
customtkinter==5.2.1
cycler==0.11.0
darkdetect==0.8.0
dill==0.3.7
Django==4.2.7
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
fonttools==4.38.0
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
kiwisolver==1.4.4
loguru==0.7.2
lxml==4.9.3
matplotlib==3.7.1
MouseInfo==0.1.3
multiprocess==0.70.14
mysqlclient==2.2.0
numpy==1.23.4
openpyxl==3.1.2
packaging==21.3
pandas==1.4.2
pathvalidate==2.5.2
Pillow==9.2.0
platformdirs==4.0.0
PyAutoGUI==0.9.54
pycryptodome==3.20.0
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
PyRect==0.2.0
PyScreeze==0.1.30
PySide2==5.15.2.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.2.0
pytz==2023.3.post1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
shiboken2==5.15.2.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
win32-setctime==1.1.0
zeep==4.2.1
zipp==3.15.0
